Weatherforecast/weather-app/weather-app/requirements.txt
requests
tkinter
python-dotenv